refactor(core): isolate diagnostics sink boundary
Keep structured log construction and GDPR redaction renderer-neutral, then pass the typed LogEntry to the existing IDB, Tauri JSONL, and development-console adapters. Preserve all logger APIs and sink behavior; this slice does not switch production authority or migrate callers.\n\nValidation: targeted Vitest 22/22, Biome, tsgo, docs metrics, native-readiness, desktop import guardrail, and diff checks passed. No new dependency or suppression was added. Local hooks remain bypassed only because pnpm reports ERR_PNPM_VERIFY_DEPS_BEFORE_RUN on the existing install; CI remains authoritative.
