+++ docs/native/CORE-MIGRATION-LEDGER.md
@@ -11,7 +11,7 @@ scope shifts — it is a living decision record, not a one-time snapshot.
 | 1 | Project schema | TS, `types.ts` (916 lines) — `StoryProject`/`Character`/`World`/`StorySection`/`Settings` | Medium — `characters`/`worlds` are typed `Character[] \| EntityState<Character, string>`, a Redux Toolkit coupling at the type level | Low | High — canonical shape everything else depends on | Low | High — every renderer needs the same shape | **1 — Wave 2 first slice** | Port the array-only member of the union to Rust structs (serde, camelCase); TS type left unchanged | Golden-master JSON fixtures decode identically in Zod and Rust |
 | 2 | Validation | TS, `services/projectImportSchema.ts` (352 lines, Zod, import-only) + `idbProjectStore.ts#validateAndFixState` (hand-written repair-on-load, separate path) | Low, but split across two divergent implementations today | Low-medium (first line of defense against malformed/malicious import files) | High | Low | High | **1 — Wave 2 first slice** | Hand-rolled Rust validation mirroring the Zod ruleset for the modeled fields; does not reconcile the Zod-vs-hand-written TS divergence | Same golden-master fixtures; accept/reject parity asserted per fixture |
 | 3 | Storage — fs backend (plain load/save) | TS, `services/fs/` (7 files, 1,253 lines) | High today (Tauri-fs-plugin coupled) | Low today — `fsCore.ts#encryptText`/`decryptText` are dead code, zero call sites | High (atomic-write semantics, compression) | Medium (large-project I/O) | High | **2 — Wave 2 first slice, narrow** | New crate proves plain JSON load/save on its own format; does not replace `projectFsStore.ts` | Round-trip identity tests inside the new crate only, not yet byte-compared against `.wsproj` |
-| 4 | Logger / diagnostics | TS, `services/logger.ts` (312 lines) — dual-sink (IDB ring buffer + Tauri JSONL), plus Rust Core proof in `crates/worldscript-diagnostics` | Medium (dual-sink dispatch remains renderer-aware) | Medium (redaction chokepoint) | Low | Low | Medium-high — small, self-contained | 3 — Wave 2 fast-follow | **In progress — Core model/redaction and TS/Rust parity are CI-proven; no authority switch** | Shared JSON fixtures are exercised independently by Rust and TS for redaction and LogEntry wire shape; sink parity and production wiring remain open |
+| 4 | Logger / diagnostics | TS, `services/logger.ts` (312 lines) — dual-sink (IDB ring buffer + Tauri JSONL), plus Rust Core proof in `crates/worldscript-diagnostics` | Medium (dual-sink dispatch remains renderer-aware) | Medium (redaction chokepoint) | Low | Low | Medium-high — small, self-contained | 3 — Wave 2 fast-follow | **In progress — Core model/redaction, TS/Rust parity, and the typed TS sink boundary are CI-proven; no authority switch** | Shared JSON fixtures are exercised independently by Rust and TS for redaction and LogEntry wire shape; browser/Tauri sink adapters now consume the portable TS LogEntry; sink parity and production wiring remain open |
 | 5 | Task-orchestration expansion | Rust, `src-tauri/src/commands/task_supervisor.rs` (bounded `text.analyze`/`text.diff`) + TS `packages/worker-bus/` (5,130 lines, full orchestration surface) | worker-bus TS side is Web-Worker-pool coupled; Rust side is renderer-neutral for the bounded proof | Low-medium | Low (task results, not persisted project state) | Medium-high | High long-term, existing Rust is a narrow stub | 4 — Wave 2 fast-follow | **Bounded task proof CI-proven in #430; full extraction not started** | Rust tests cover structured failure, input bounds, deterministic analysis, diff parity, and the v1.0.0 request/result contract; no broad worker-bus authority switch |
 | 6 | Storage — IDB backend (all encryption) | TS, `services/storage/` (18 files, 5,363 lines) — mature AES-256-GCM, ADR-0018 "Accepted and implemented" | High — IndexedDB is browser-only, not portable to Qt/GPUI as-is | High, mature | High | Medium | Low as literally written; high as a design reference | **Deferred to Wave 3-4** | None in Wave 2. ADR-0018's 6 invariants are required reading for Wave 3 crypto design | None in Wave 2 |
 | 7 | `features/project/` domain logic | TS, `features/project/` (24 files, 2,114 lines) — real logic concentrated in `thunks/` + `projectSelectors.ts` (~450-500 lines); `reducers/` (11 files) is CRUD bookkeeping | High — Redux-store-shape/dispatch bound; `reducers/` stays TS-side permanently | Low | Medium (import/restore orchestration) | Low-medium | Medium (only the thunks/selectors subset) | Deferred | Candidate after the schema crate is proven; only thunks/selectors, never `reducers/` | Not started |
@@ -40,10 +40,12 @@ scope shifts — it is a living decision record, not a one-time snapshot.
   entry without a matching entity, an entity absent from `ids`, or duplicate IDs; it never silently
   drops, synthesizes, or last-write-wins. Valid input preserves the declared `ids` order and
   entity IDs in the boundary fixtures. This proves the contract, not a production authority switch.
-- **Row 4 has a narrow CI-proven Core proof.** `crates/worldscript-diagnostics` owns only the
-  typed structured-log model and the existing top-level key-redaction semantics; PR #437 proves
-  the redaction contract independently against the TypeScript implementation. IDB, Tauri JSONL,
-  and development-console sinks remain TS-side; no production caller or authority switch is claimed.
+- **Row 4 has a narrow CI-proven Core proof and a typed sink boundary.** `crates/worldscript-diagnostics`
+  owns only the typed structured-log model and the existing top-level key-redaction semantics;
+  PR #437 proves the redaction contract independently against the TypeScript implementation. The
+  browser/Tauri sink adapters now receive the portable TypeScript `LogEntry` from
+  `services/diagnostics/logSinks.ts`; IDB, Tauri JSONL, and development-console ownership remains
+  renderer-specific, and no production caller or authority switch is claimed.
 - Row 5 is already locally and CI-proven as a bounded Tauri supervisor proof (`text.analyze` and
   `text.diff`, with structured failure semantics); full worker-bus extraction and G1 completion remain open.
